Resolve issue #2: CLI argument parsing
This commit adds command line argument option parsing using the `yargs`
library, which generates nice usage/help text and appears to be pretty
flexible. It's also pirate-themed, and my boat is full of limes.

What shall we do with a drunken soldier?
Put him in the guard room 'til he gets sober.
Put him in the scuppers with a hosepipe on him.
Pull out the plug and wet him all over.
Tie him to the taffrail when she's yardarm under.
Heave him by the leg in a runnin' bowline.
Scrape the hair off his chest with a hoop-iron razor.
Give 'im a dose of salt and water.
Stick on his back a mustard plaster.
Keep him there and make 'im bail 'er.
Give 'im a taste of the bosun's rope-end.
Early in the morning.