Resolve issue #11
This commit looks bigger than it really is because it includes
some moving around of blocks and some minimal refactoring. The
results appear to match the UI, but this could use more testing.

For some reason, this API endpoint also doesn't accept a `limit`
parameter. I tried to report the issue, but their bug reporting
page is so catastrophically broken (it doesn't even render correctly)
that ascertaining a valid security contact is impossible.

Document and move on.