Actually send the JST token
For some reason, this appears to be required for some Parler sessions but
not forever. There isn't really a good explanation for this that I can find,
other than there may be a server-side defect. To be a good citizen and avoid
doing anything that could potentially degrade service or cause damage, send
this short-lived token exactly as the web client does.