Toward issue #22 and issue #18
Refactor the file I/O into the Output class, and add a generic
promisification wrapper to deal with earlier versions of Node.

This needs to be cleaned up before it's shipped, but in theory the
`Output` class can provide browser-specific I/O methods when the
client is running in a browser or on another Javascript engine, and
not on Node.js.