Add link and creator expansion
Some work still needs to be done re: formatting of reparented
results. For instance, it'd probably be nice if the links were an object
keyed by link ID, rather than an array of objects each containing a `link`
descriptor object. For now, keep the regular (if unwieldly in some cases)
structure.