Resolve regression in issue #36
Refactoring in the IO subsystem (intended to make it easier to run
the client inside of an ordinary web browser in the near future)
led to the `init` subcommand attempting to use the *base* I/O
class instead of the Node-specific subclass. The base class leaves
`read_file` and `write_file` unimplemented, because there exists
no method of reading and writing files that's browser-portable.

Since this only runs under Node currently, use `IOH.Node`. Doh.