Toward #37: Disable automatic boolean negation
As mentioned in https://github.com/yargs/yargs/issues/1011, the `yargs`
argument parsing library contains code that attempts to intelligently
handle negation of boolean arguments via a `--no-` prefix. Disable
this behavior; it's undesirable given the set of options we currently
provide, and easy enough to deal with manually should the need arise.

This issue was causing one or more long-form options to be incorrectly
ignored by the CLI interface.