Fix harmless fencepost error; conjure OCD
This does a little bit of cleanup related to issue #38; nothing major.
I've tested this with an unreliable connection and it seems to work
pretty well. I need to find a good way to simulate non-2XX HTTP
responses in order to unit test this (see #10).

Thanks again to @milesmcc for getting proper request error handling
and retries going.