Fix argument-parsing issues, toward issue #37
This should now issue HTTPS requests with the proper `limit` parameter,
although most Parler API endpoints appear to ignore the option now.
Which I guess is good; it's a baby step toward learning that they've
made a very bad decision re: their choice of datastore.

There are probably more argument parsing issues; I was expecting a
behavior from the `yargs` argument-parsing library that's documented
but doesn't currently seem to work.

If Parler's actually going to be a thing in 2021, it's probably time
to prioritize issue #10, and get this stuff unit tested to the extent
possible.